ugly but funcitonal ui
